+++ pylibfranka/include/pylibfranka.h
@@ -69,6 +69,13 @@ class PyRobot {
   auto startCartesianPoseControl(franka::ControllerMode controller_mode)
       -> std::unique_ptr<franka::ActiveControlBase>;
 
+  /**
+   * Starts the Cartesian velocity control mode.
+   * @param control_type The type of controller to use (JointImpedance or CartesianImpedance).
+   */
+  auto startCartesianVelocityControl(franka::ControllerMode controller_mode)
+      -> std::unique_ptr<franka::ActiveControlBase>;
+
   // Configuration methods
   void setCollisionBehavior(const std::array<double, 7>& lower_torque_thresholds,
                             const std::array<double, 7>& upper_torque_thresholds,